Refactor out elasticsearch placeholders to super
So when we have more prometheus metric data getters they can use the
same function

As it is a pure function this could probably not be an inherited class
/shrug
